# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me/files/{id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Get file object (not supported)",
                "description": "This endpoint is not supported. Use POST to upload files.",
                "operationId": "getFileObject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File object ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Method not allowed - use POST to upload files"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/files": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Upload a file",
                "description": "Uploads a file and returns a FileObject that can be referenced in bookmarks (mainImage, archive).",
                "operationId": "uploadFile",
                "requestBody": {
                    "description": "File to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FileObject"
                                },
                                "examples": {
                                    "uploaded_file": {
                                        "summary": "Successfully uploaded file",
                                        "value": {
                                            "contentUrl": "https://bookmarkhive.test/storage/files/abc123.jpg",
                                            "size": 102400,
                                            "mime": "image/jpeg",
                                            "@iri": "https://bookmarkhive.test/users/me/files/01234567-89ab-cdef-0123-456789abcdef"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "422": {
                        "description": "Validation error - invalid file",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_file": {
                                        "summary": "Invalid file validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "description": "Error object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "description": "HTTP status code",
                                "type": "integer"
                            },
                            "message": {
                                "description": "Error message",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "FileObject": {
                "description": "File object representing an uploaded file",
                "properties": {
                    "contentUrl": {
                        "description": "URL to access the file content",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "size": {
                        "description": "File size in bytes",
                        "type": "integer"
                    },
                    "mime": {
                        "description": "MIME type of the file",
                        "type": "string"
                    },
                    "@iri": {
                        "description": "IRI of the file object resource",
                        "type": "string",
                        "format": "iri"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}